parlay.py
Simple app that creates option chain based off moneyline odds on options implied probabilities.  
Typically, in the money options that are a 50/50 proposition, can be estimated at +100 as a moneyline bet in American Odds. 

Here we are calculating only 50% implied probability and 30% implied probability.  

We can then stack the events/outcomes and guess what our returns would be.